Calibrate_Board
I am working on a script that will image and calibrate the IP cameras
with the OpenCV checkerboard with just a few steps.  This works with the
SV3C cameras from Amazon.  This is a work in progress.  I do not have a
way to export the calibration data yet.